+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-16 20:05:19
-// Commit: feat(core): integrate custom telemetry logger in core
+// Updated at: 2026-05-17 14:46:52
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '1211';
-  final String lastMessage = 'feat(core): integrate custom telemetry logger in core';
-  final DateTime timestamp = DateTime.parse('2026-05-16T20:05:19+05:00');
+  final String commitRef = '1223';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-05-17T14:46:52+05:00');
 }