+++ CHANGELOG.md
@@ -1224,3 +1224,4 @@
 - [2026-05-17 10:22] refactor(i18n): extract reusable atomic widgets from i18n
 - [2026-05-17 10:27] feat(cache): add async cache stream provider
 - [2026-05-17 14:19] refactor(navigation): optimize data flow between domain and data layers
+- [2026-05-17 14:46] feat(core): add bi-directional event bus handler for core